@@ -1,15 +1,16 @@
 [project]
 name = "aoc"
-version = "0.3.0"
+version = "0.4.0"
 description = "Advent of Code tools"
 readme = "README.md"
 license = "MIT"
 license-files = ["LICENSE"]
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 dependencies = [
     "click>=8.0",
     "beautifulsoup4>=4.12",
-    "requests>=2.30"
+    "requests>=2.30",
+    "toml>=0.10",
 ]
 
 [project.optional-dependencies]
